+++ SupremacyClient/SupremacyClient.csproj
@@ -193,9 +193,6 @@
       <HintPath>..\Common\ShaderEffectLibrary.dll</HintPath>
       <Private>False</Private>
     </Reference>
-    <Reference Include="SoftUni.Wintellect.PowerCollections, Version=1.1.5733.22550, Culture=neutral, processorArchitecture=MSIL">
-      <HintPath>..\packages\SoftUni.Wintellect.PowerCollections.1.1.5733.22550\lib\net20\SoftUni.Wintellect.PowerCollections.dll</HintPath>
-    </Reference>
     <Reference Include="System" />
     <Reference Include="System.configuration" />
     <Reference Include="System.Core">
