+++ SupremacyClient/packages.config
@@ -5,5 +5,4 @@
   <package id="Microsoft.Expression.Drawing" version="3.0.0" targetFramework="net40" requireReinstallation="true" />
   <package id="MicrosoftExpressionInteractions" version="3.0.40218.0" targetFramework="net40" requireReinstallation="true" />
   <package id="PowerThreading" version="20120420.0" targetFramework="net40" />
-  <package id="SoftUni.Wintellect.PowerCollections" version="1.1.5733.22550" targetFramework="net40" />
 </packages>
