+++ SupremacyClientComponents/ImageCache.cs
@@ -9,127 +9,10 @@
 
 using System;
 using System.Windows.Media.Imaging;
-
-using Supremacy.Collections;
-
 using TvdP.Collections;
 
 namespace Supremacy.Client
 {
-    public abstract class Cache<T> where T : class
-    {
-        private readonly CacheEntryCollection<T> _cache;
-
-        protected Cache()
-        {
-            _cache = new CacheEntryCollection<T>();
-        }
-
-        public T this[string key]
-        {
-            get
-            {
-                if (_cache.Contains(key))
-                {
-                    if (_cache[key].Reference.IsAlive)
-                    {
-                        return _cache[key].Reference.Target as T;
-                    }
-                    _cache.Remove(key);
-                    return Get(key);
-                }
-                return null;
-            }
-            set
-            {
-                if (value != null)
-                {
-                    if (_cache.Contains(key))
-                    {
-                        var reference = _cache[key].Reference;
-                        if (!reference.IsAlive || (reference.Target != value))
-                            _cache.Add(new CacheEntry(key, value));
-                    }
-                    else
-                    {
-                        _cache.Add(new CacheEntry(key, value));
-                    }
-                }
-                else if (_cache.Contains(key))
-                {
-                    _cache.Remove(key);
-                }
-            }
-        }
-
-        public T Get(Uri uri)
-        {
-            if (uri == null)
-                throw new ArgumentNullException("uri");
-            return Get(uri.OriginalString);
-        }
-
-        public T Get(string uri)
-        {
-            if (uri == null)
-                throw new ArgumentNullException("uri");
-            if (uri == null)
-            {
-                throw new ArgumentNullException("uri");
-            }
-            if (_cache.Contains(uri))
-            {
-                if (_cache[uri].Reference.IsAlive)
-                {
-                    return _cache[uri].Reference.Target as T;
-                }
-                _cache.Remove(uri);
-            }
-            try
-            {
-                var item = Load(uri);
-                if (item != null)
-                {
-                    this[uri] = item;
-                }
-                return item;
-            }
-            catch
-            {
-                return null;
-            }
-        }
-
-        protected abstract T Load(string uri);
-
-        public struct CacheEntry
-        {
-            private readonly string _key;
-            private readonly WeakReference _reference;
-
-            public string Key
-            {
-                get { return _key; }
-            }
-
-            public WeakReference Reference
-            {
-                get { return _reference; }
-            }
-
-            public CacheEntry(string key, T target)
-            {
-                _key = key;
-                _reference = new WeakReference(target);
-            }
-        }
-    }
-
-    public sealed class CacheEntryCollection<T> : KeyedCollectionBase<string, Cache<T>.CacheEntry> where T : class
-    {
-        public CacheEntryCollection() : base(o => o.Key) {}
-    }
-
     public sealed class ImageCache
     {
         private static readonly ImageCache s_current;
