+++ SupremacyClientComponents/packages.config
@@ -2,5 +2,4 @@
 <packages>
   <package id="AvalonLibrary" version="2.3.0.0" targetFramework="net40" />
   <package id="log4net" version="2.0.8" targetFramework="net40" requireReinstallation="true" />
-  <package id="SoftUni.Wintellect.PowerCollections" version="1.1.5733.22550" targetFramework="net40" />
 </packages>
