+++ SupremacyCore/Collections/StackExtensions.cs
@@ -5,6 +5,19 @@ namespace Supremacy.Collections
 {
     public static class StackExtensions
     {
+        public static IStack<T> Push<T>(this IStack<T> s, T t)
+        {
+            return ImmutableStack<T>.Push(t, s);
+        }
+
+        static public IStack<T> Reverse<T>(this IStack<T> stack)
+        {
+            IStack<T> r = ImmutableStack<T>.Empty;
+            for (IStack<T> f = stack; !f.IsEmpty; f = f.Pop())
+                r = r.Push(f.Peek());
+            return r;
+        }
+
         public static bool TryPop<T>(this Stack<T> stack, out T item)
         {
             if (stack == null)
