+++ SupremacyCore/packages.config
@@ -2,6 +2,5 @@
 <packages>
   <package id="C5" version="2.5.3" targetFramework="net40" requireReinstallation="true" />
   <package id="log4net" version="2.0.8" targetFramework="net40" requireReinstallation="true" />
-  <package id="SoftUni.Wintellect.PowerCollections" version="1.1.5733.22550" targetFramework="net40" />
   <package id="System.ValueTuple" version="4.5.0" targetFramework="net462" />
 </packages>
