+++ SupremacyUI/SupremacyUI.csproj
@@ -79,9 +79,6 @@
       <HintPath>..\Common\ShaderEffectLibrary.dll</HintPath>
       <Private>False</Private>
     </Reference>
-    <Reference Include="SoftUni.Wintellect.PowerCollections, Version=1.1.5733.22550, Culture=neutral, processorArchitecture=MSIL">
-      <HintPath>..\packages\SoftUni.Wintellect.PowerCollections.1.1.5733.22550\lib\net20\SoftUni.Wintellect.PowerCollections.dll</HintPath>
-    </Reference>
     <Reference Include="System" />
     <Reference Include="System.Core">
       <RequiredTargetFramework>3.5</RequiredTargetFramework>
